refactor: range index sorted
