--- src/cid-index-sorted/codec.js
@@ -1,4 +1,0 @@
-/**
- * Codec for the CIDIndexSorted format.
- */
-export const CID_INDEX_SORTED_CODEC = 0x0403
