--- src/cid-index-sorted/index.js
@@ -1,2 +1,0 @@
-export * as CIDIndexSortedReader from './reader.js'
-export * as CIDIndexSortedWriter from './writer.js'
