+++ src/index-sorted/README.md
@@ -2,11 +2,11 @@
 
 ```
 varint codec
-    uint32 number of "width" groups
+uint32 number of "width" groups
+    (for each)
+    uint32 group "width"
+    uint64 group length
         (for each)
-        uint32 group "width"
-        uint64 group length
-            (for each)
-            digest
-            uint64 offset
+        digest
+        uint64 offset
 ```
