+++ src/range-index-sorted/api.js
