+++ src/range-index-sorted/codec.js
@@ -1,0 +1,4 @@
+/**
+ * Codec for the RangeIndexSorted format.
+ */
+export const RANGE_INDEX_SORTED_CODEC = 0x0403
