+++ src/range-index-sorted/index.js
@@ -1,0 +1,2 @@
+export * as RangeIndexSortedReader from './reader.js'
+export * as RangeIndexSortedWriter from './writer.js'
